refactor: header search elements now in ui dir
