+++ src/Home/Header/SearchButton/index.js
@@ -2,12 +2,12 @@ import React, { Component } from "react";
 import { withRouter } from "react-router-dom";
 
 class SearchButton extends Component {
-  goSearch = () => {
+  gotoSearch = () => {
     this.props.history.push("/search");
   };
   render() {
     return (
-      <button onClick={this.goSearch} className="header__search-button">
+      <button onClick={this.gotoSearch} className="header__search-button">
         Найти билеты
       </button>
     );
