+++ src/Home/Header/styles.sass
@@ -61,7 +61,7 @@
     &__from-city
       border-top-left-radius: 4px
       border-top-right-radius: 4px
-      background: #FFFFFF url('./SearchFilters/arrow.svg') no-repeat 95% 48%
+      background: #FFFFFF url('./../../ui/SearchFilters/arrow.svg') no-repeat 95% 48%
       padding: 16px
       margin: 2px 0
       +tablet
@@ -127,7 +127,7 @@
         min-width: 0
       &__wrapper
         width: 50%
-        background: #FFFFFF url('./SearchFilters/calendar.svg') no-repeat 90% 48% 
+        background: #FFFFFF url('./../../ui/SearchFilters/calendar.svg') no-repeat 90% 48% 
         padding: 16px
         &:nth-child(1)
           +tablet
@@ -168,7 +168,7 @@
         font-size: 16px
         width: 100%
   &__search-button
-    background: #FF9241 url('./SearchButton/aero.svg') no-repeat 80% 50%
+    background: #FF9241 url('./../../ui/SearchButton/aero.svg') no-repeat 80% 50%
     border-radius: 4px
     padding: 14px 88px 16px 55px
     font-family: Roboto
